Update sleep chart and alerts
