+++ template/html.py
@@ -22,11 +22,11 @@ def head():
 def header():
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div id="top" class='intro_section'>
-        <p><img class='intro_logo' src='data:image/png;base64,""" + st.session_state.logo_clife_white + """'/></p>
+        <p><img class='intro_logo' src='data:image/png;base64,{st.session_state.logo_clife_white}'/></p>
         <center>
-        <p class='intro_style'><b>""" + translate["header_title"] + """</b></p>
+        <p class='intro_style'><b>{translate["header_title"]}</b></p>
         </center>
     </div>
     """
@@ -50,9 +50,9 @@ def language_label():
     translate = st.session_state.translate
     
     
-    html = """
+    html = f"""
     <div>
-        <p class='language_label'><b>""" + translate["language_label"] + """</b></p>
+        <p class='language_label'><b>{translate["language_label"]}</b></p>
     </div>
     """
     
@@ -62,36 +62,35 @@ def profile():
     
     username = st.session_state.username
     
-    html = """
+    html = f"""
     <div class="user_icon_section">
         <p>    
-            <img class="user_icon" src='data:image/png;base64,""" + st.session_state.user_icon + """'/>
+            <img class="user_icon" src='data:image/png;base64,{st.session_state.user_icon}'/>
             <br>
         </p>
-        <p>""" + username + """</p>
+        <p>{username}</p>
     </div>
     """
     
     return html
 
-# def button_scroll_to_top():
-#     html = """
+#/ def button_scroll_to_top():
+#     html = f"""
 #     <div id="scrollUp" class="scrollUp">
-#         <a href="#top"><img src='data:image/png;base64,""" + st.session_state.to_top + """'/></a>
+#         <a href="#top"><img src='data:image/png;base64,{st.session_state.to_top}'/></a>
 #     </div>
-#     """
-    
+#     """  
 #     return html
 
 def menu_overview():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="menu">
         <p>
             <a href="#overview">
-                <span>""" + translate["overview"] + """</span>
+                <span>{translate["overview"]}</span>
             </a>
         </p>
     </div>
@@ -102,11 +101,11 @@ def menu_overview():
 def menu_smart_textile_raw_data():
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="menu">
         <p>
             <a href="#smart_textile_raw_data">
-                <span>""" + translate["smart_textile_raw_data"] + """</span>
+                <span>{translate["smart_textile_raw_data"]}</span>
             </a>
         </p>
     </div>
@@ -118,11 +117,11 @@ def menu_health_indicators():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="menu">
         <p>
             <a href="#health_indicators">
-                <span>""" + translate["health_indicators"] + """</span>
+                <span>{translate["health_indicators"]}</span>
             </a>
         </p>
     </div>
@@ -134,11 +133,11 @@ def menu_data_report():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="menu">
         <p>
             <a href="#data_report">
-                <span>""" + translate["data_report"] + """</span>
+                <span>{translate["data_report"]}</span>
             </a>
         </p>
     </div>
@@ -150,11 +149,11 @@ def menu_definitions():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="menu">
         <p>
             <a href="#definitions">
-                <span>""" + translate["definitions"] + """</span>
+                <span>{translate["definitions"]}</span>
             </a>
         </p>
     </div>
@@ -166,9 +165,9 @@ def overview_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
         <div id="overview" class="main_title">
-            <p>""" + translate["overview"] + """</p>
+            <p>{translate["overview"]}</p>
         </div>
         <div>
     """
@@ -203,121 +202,121 @@ def overview_data_collection():
     col_val         = "col-5"
     class_val       = "indicator_value"
     
-    html = """
+    html = f"""
         <div class="second_title">
-            <p>""" + translate["data_collection"] + """</p>
+            <p>{translate["data_collection"]}</p>
         </div>
         <div class="row">
-            <div class='""" + col_section + """'>
+            <div class='{col_section}'>
                 <div class='overview_section'>
                     <div class="row">
-                        <div class='""" + col_title + """'>
-                            <p class='indicator_name'>""" + translate["smart_textile"] + """</p>
+                        <div class='{col_title}'>
+                            <p class='indicator_name'>{translate["smart_textile"]}</p>
                         </div>
                     </div>
                     <div class="row overview_sub_section_smart_textile">
-                        <div class='""" + col_logo + """'>
-                            <img class=device_icon src='data:image/png;base64,""" + st.session_state.tshirt_right + """'/>
+                        <div class='{col_logo}'>
+                            <img class=device_icon src='data:image/png;base64,{st.session_state.tshirt_right}'/>
                         </div>
-                        <div class='""" + col_data + """'>
+                        <div class='{col_data}'>
                             <div class="row data_collection">
-                                <div class='""" + col_duration + """'>
-                                    <p>""" + translate["total_data"] + """ &ensp; <span class='collect_duration'>""" +\
-                                        str(chronolife_duration) + """</span></p> 
+                                <div class='{col_duration}'>
+                                    <p>{translate["total_data"]} &ensp; <span class='collect_duration'>
+                                        {str(chronolife_duration)}</span></p> 
                                 </div>
-                                <div class='""" + col_icon + """'>
-                                    <img class=miniicon src='data:image/png;base64,""" + st.session_state.night + """'/> 
+                                <div class='{col_icon}'>
+                                    <img class=miniicon src='data:image/png;base64,{st.session_state.night}'/> 
                                 </div>
-                                <div class='""" + col_key + """'>
-                                    <p>""" + translate["night"] + """</p>
+                                <div class='{col_key}'>
+                                    <p>{translate["night"]}</p>
                                 </div> 
-                                <div class='""" + col_val + """'>
-                                    <p class='""" + class_val + """'>""" + str(chronolife_duration_night) + """</p>
+                                <div class='{col_val}'>
+                                    <p class='{class_val}'>{str(chronolife_duration_night)}</p>
                                 </div> 
-                                <div class='""" + col_icon + """'>
-                                    <img class=miniicon src='data:image/png;base64,""" + st.session_state.day + """'/> 
+                                <div class='{col_icon}'>
+                                    <img class=miniicon src='data:image/png;base64,{st.session_state.day}'/> 
                                 </div>
-                                <div class='""" + col_key + """'>
-                                    <p>""" + translate["day"] + """</p>
+                                <div class='{col_key}'>
+                                    <p>{translate["day"]}</p>
                                 </div> 
-                                <div class='""" + col_val + """'>
-                                    <p class='""" + class_val + """'>""" + str(chronolife_duration_day)+ """</p>
+                                <div class='{col_val}'>
+                                    <p class='{class_val}'>{str(chronolife_duration_day)}</p>
                                 </div> 
-                                <div class='""" + col_icon + """'>
-                                    <img class=miniicon src='data:image/png;base64,""" + st.session_state.rest + """'/> 
+                                <div class='{col_icon}'>
+                                    <img class=miniicon src='data:image/png;base64,{st.session_state.rest}'/> 
                                 </div>
-                                <div class='""" + col_key + """'>
-                                    <p>""" + translate["rest"] + """</p>
+                                <div class='{col_key}'>
+                                    <p>{translate["rest"]}</p>
                                 </div> 
-                                <div class='""" + col_val + """'>
-                                    <p class='""" + class_val + """'>""" + str(chronolife_duration_rest) + """</p>
+                                <div class='{col_val}'>
+                                    <p class='{class_val}'>{str(chronolife_duration_rest)}</p>
                                 </div> 
-                                <div class='""" + col_icon + """'>
-                                    <img class=miniicon src='data:image/png;base64,""" + st.session_state.activity + """'/> 
+                                <div class='{col_icon}'>
+                                    <img class=miniicon src='data:image/png;base64,{st.session_state.activity}'/> 
                                 </div>
-                                <div class='""" + col_key + """'>
-                                    <p>""" + translate["activity"] + """</p>
+                                <div class='{col_key}'>
+                                    <p>{translate["activity"]}</p>
                                 </div> 
-                                <div class='""" + col_val + """'>
-                                    <p class='""" + class_val + """'>""" + str(chronolife_duration_activity) + """</p>
+                                <div class='{col_val}'>
+                                    <p class='{class_val}'>{str(chronolife_duration_activity)}</p>
                                 </div> 
                             </div>
                         </div>
                     </div>
                 </div>
             </div>
-            <div class='""" + col_section + """'>
+            <div class='{col_section}'>
                 <div class='overview_section'> 
                     <div class="row">
-                        <div class='""" + col_title + """'>
-                            <p class='indicator_name'>""" + translate["garmin_watch"] + """</p>
+                        <div class='{col_title}'>
+                            <p class='indicator_name'>{translate["garmin_watch"]}</p>
                         </div>
                     </div>
                     <div class="row overview_sub_section_garmin">
-                        <div class='""" + col_logo + """'>
-                            <img class=device_icon src='data:image/png;base64,""" + st.session_state.garrmin + """'/>
+                        <div class='{col_logo}'>
+                            <img class=device_icon src='data:image/png;base64,{st.session_state.garrmin}'/>
                         </div>
-                        <div class='""" + col_data + """'>
+                        <div class='{col_data}'>
                             <div class="row data_collection">
-                                <div class='""" + col_duration + """'>
-                                    <p>""" + translate["total_data"] + """ &ensp; <span class='collect_duration'>""" +\
-                                        str(garmin_duration) + """</span></p> 
+                                <div class='{col_duration}'>
+                                    <p>{translate["total_data"]} &ensp; <span class='collect_duration'>
+                                        {str(garmin_duration)}</span></p> 
                                 </div>
-                                <div class='""" + col_icon + """'>
-                                    <img class=miniicon src='data:image/png;base64,""" + st.session_state.night + """'/> 
+                                <div class='{col_icon}'>
+                                    <img class=miniicon src='data:image/png;base64,{st.session_state.night}'/> 
                                 </div>
-                                <div class='""" + col_key + """'>
-                                    <p>""" + translate["night"] + """</p>
+                                <div class='{col_key}'>
+                                    <p>{translate["night"]}</p>
                                 </div> 
-                                <div class='""" + col_val + """'>
-                                    <p class='""" + class_val + """'>""" + str(garmin_duration_night) + """</p>
+                                <div class='{col_val}'>
+                                    <p class='{class_val}'>{str(garmin_duration_night)}</p>
                                 </div> 
-                                <div class='""" + col_icon + """'>
-                                    <img class=miniicon src='data:image/png;base64,""" + st.session_state.day + """'/> 
+                                <div class='{col_icon}'>
+                                    <img class=miniicon src='data:image/png;base64,{st.session_state.day}'/> 
                                 </div>
-                                <div class='""" + col_key + """'>
-                                    <p>""" + translate["day"] + """</p>
+                                <div class='{col_key}'>
+                                    <p>{translate["day"]}</p>
                                 </div> 
-                                <div class='""" + col_val + """'>
-                                    <p class='""" + class_val + """'>""" + str(garmin_duration_day) + """</p>
+                                <div class='{col_val}'>
+                                    <p class='{class_val}'>{str(garmin_duration_day)}</p>
                                 </div> 
-                                <div class='""" + col_icon + """'>
-                                    <img class=miniicon src='data:image/png;base64,""" + st.session_state.rest + """'/> 
+                                <div class='{col_icon}'>
+                                    <img class=miniicon src='data:image/png;base64,{st.session_state.rest}'/> 
                                 </div>
-                                <div class='""" + col_key + """'>
-                                    <p>""" + translate["rest"] + """</p>
+                                <div class='{col_key}'>
+                                    <p>{translate["rest"]}</p>
                                 </div> 
-                                <div class='""" + col_val + """'>
-                                    <p class='""" + class_val + """'>""" + str(garmin_duration_rest) + """</p>
+                                <div class='{col_val}'>
+                                    <p class='{class_val}'>{str(garmin_duration_rest)}</p>
                                 </div> 
-                                <div class='""" + col_icon + """'>
-                                    <img class=miniicon src='data:image/png;base64,""" + st.session_state.activity + """'/> 
+                                <div class='{col_icon}'>
+                                    <img class=miniicon src='data:image/png;base64,{st.session_state.activity}'/> 
                                 </div>
-                                <div class='""" + col_key + """'>
-                                    <p>""" + translate["activity"] + """</p>
+                                <div class='{col_key}'>
+                                    <p>{translate["activity"]}</p>
                                 </div> 
-                                <div class='""" + col_val + """'>
-                                    <p class='""" + class_val + """'>""" + str(garmin_duration_activity) + """</p>
+                                <div class='{col_val}'>
+                                    <p class='{class_val}'>{str(garmin_duration_activity)}</p>
                                 </div> 
                             </div>
                         </div>
@@ -410,356 +409,356 @@ def overview_health_indicators():
     col_val     = "col-6"
     class_val   = "indicator_value"
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <p>""" + translate["health_indicators_overview"] + """</p>
+        <p>{translate["health_indicators_overview"]}</p>
     </div>
     <div class="row">
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='overview_health_section'> 
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.heart_icon + """'/> 
-                        <p class='indicator_name'>""" + translate["cardiology"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.heart_icon}'/> 
+                        <p class='indicator_name'>{translate["cardiology"]}</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["tachycardia"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["tachycardia"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p><img class=miniicon src='data:image/png;base64,""" + st.session_state.tachycardia_alert_icon + """'/></p>
+                    <div class='{col_val}'>
+                        <p><img class=miniicon src='data:image/png;base64,{st.session_state.tachycardia_alert_icon}'/></p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["bradycardia"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["bradycardia"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p><img class=miniicon src='data:image/png;base64,""" + st.session_state.bradycardia_alert_icon + """'/> </p>
+                    <div class='{col_val}'>
+                        <p><img class=miniicon src='data:image/png;base64,{st.session_state.bradycardia_alert_icon}'/> </p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["qt_anomaly"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["qt_anomaly"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p><img class=miniicon src='data:image/png;base64,""" + st.session_state.qt_alert_icon + """'/> </p>
+                    <div class='{col_val}'>
+                        <p><img class=miniicon src='data:image/png;base64,{st.session_state.qt_alert_icon}'/> </p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["hr_high"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["hr_high"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(bpm_high) + """ bpm</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(bpm_high)} bpm</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["hr_resting"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["hr_resting"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(bpm_rest) + """ bpm</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(bpm_rest)} bpm</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["hrv_resting"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["hrv_resting"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(hrv_rest) + """ ms</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(hrv_rest)} ms</p>
                     </div>
                 </div>
             </div>
         </div>
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='overview_health_section'> 
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.breath_icon + """'/> 
-                        <p class='indicator_name'>""" + translate["respiratory"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.breath_icon}'/> 
+                        <p class='indicator_name'>{translate["respiratory"]}</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["tachypnea"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["tachypnea"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p><img class=miniicon src='data:image/png;base64,""" + st.session_state.tachypnea_alert_icon + """'/> </p>
+                    <div class='{col_val}'>
+                        <p><img class=miniicon src='data:image/png;base64,{st.session_state.tachypnea_alert_icon}'/> </p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["bradypnea"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["bradypnea"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p><img class=miniicon src='data:image/png;base64,""" + st.session_state.bradypnea_alert_icon + """'/> </p>
+                    <div class='{col_val}'>
+                        <p><img class=miniicon src='data:image/png;base64,{st.session_state.bradypnea_alert_icon}'/> </p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["br_high"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["br_high"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(brpm_high) + """ brpm</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(brpm_high)} brpm</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["br_resting"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["br_resting"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(brpm_rest) + """ brpm</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(brpm_rest)} brpm</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["brv_resting"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["brv_resting"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(brv_rest) + """ s</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(brv_rest)} s</p>
                         <br>
                     </div>
                 </div>
             </div>
         </div>
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='overview_health_section'> 
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.pulseox_icon + """'/> 
-                        <p class='indicator_name'>""" + translate["pulseox"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.pulseox_icon}'/> 
+                        <p class='indicator_name'>{translate["pulseox"]}</p>
                     </div>
                     <div class="col-6">
-                            <img class=coloricon src='data:image/png;base64,""" + st.session_state.spo2_green + """'/>
+                            <img class=coloricon src='data:image/png;base64,{st.session_state.spo2_green}'/>
                             <p>90 - 100 %</p>
-                            <img class=coloricon src='data:image/png;base64,""" + st.session_state.spo2_yellow + """'/>
+                            <img class=coloricon src='data:image/png;base64,{st.session_state.spo2_yellow}'/>
                             <p>80 - 89 %</p>
-                            <img class=coloricon src='data:image/png;base64,""" + st.session_state.spo2_orange + """'/>
+                            <img class=coloricon src='data:image/png;base64,{st.session_state.spo2_orange}'/>
                             <p>70 - 79 %</p>
-                            <img class=coloricon src='data:image/png;base64,""" + st.session_state.spo2_red + """'/>
+                            <img class=coloricon src='data:image/png;base64,{st.session_state.spo2_red}'/>
                             <p>< 70 %</p>
                     </div>
                     <div class="col-6">
-                        <img class=donut src='data:image/png;base64,""" + st.session_state.spo2_donut + """'/> 
+                        <img class=donut src='data:image/png;base64,{st.session_state.spo2_donut}'/> 
                     </div>
                 </div>
             </div>
         </div>
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='overview_health_section'> 
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.steps_icon + """'/> 
-                        <p class='indicator_name'>""" + translate["steps"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.steps_icon}'/> 
+                        <p class='indicator_name'>{translate["steps"]}</p>
                     </div>
                     <div class="col-6">
                         <div class="row">
                             <div class="col-12">
-                                <p>""" + translate["number"] + """</p>
+                                <p>{translate["number"]}</p>
                             </div>    
                             <div class="col-12">
-                                <p class='steps_number'>""" + str(steps_number) + """</p>
+                                <p class='steps_number'>{str(steps_number)}</p>
                             </div>
                             <div class="col-12">
-                                <p>""" + translate["goal"] + """</p>
+                                <p>{translate["goal"]}</p>
                             </div>
                             <div class="col-12">
-                                <p><b>""" + str(goal) + """</b></p>
+                                <p><b>{str(goal)}</b></p>
                             </div>
                             <div class="col-12">
-                                <p>""" + translate["distance"] + """</p>
+                                <p>{translate["distance"]}</p>
                             </div>
                             <div class="col-12">
-                                <p><b>""" + str(distance) + """ km</b></p>
+                                <p><b>{str(distance)} m</b></p>
                             </div>
                         </div>
                     </div>
                     <div class="col-6">
                         <center>
-                            <img class=donut src='data:image/png;base64,""" + st.session_state.steps_donut + """'/> 
+                            <img class=donut src='data:image/png;base64,{st.session_state.steps_donut}'/> 
                         </center>
                     </div>
                 </div>
             </div>
         </div>
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='overview_health_section'> 
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.stress_icon + """'/> 
-                        <p class='indicator_name'>""" + translate["stress"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.stress_icon}'/> 
+                        <p class='indicator_name'>{translate["stress"]}</p>
                     </div>
                     <div class="col-7">
                         <div class="row">
                             <div class="col-12">
-                                <img class=coloricon src='data:image/png;base64,""" + st.session_state.stress_rest + """'/> 
+                                <img class=coloricon src='data:image/png;base64,{st.session_state.stress_rest}'/> 
                                 <p>
-                                    """ + translate["rest"] + """ <br>
-                                    <span class='""" + class_val + """'>""" + str(stress_duration_rest) + """ (""" + str(stress_percentage_rest) + """%)</span>
+                                    {translate["rest"]} <br>
+                                    <span class='{class_val}'>{str(stress_duration_rest)} ({str(stress_percentage_rest)}%)</span>
                                 </p>
                             </div>
                             <div class="col-12">
-                                <img class=coloricon src='data:image/png;base64,""" + st.session_state.stress_low + """'/>
+                                <img class=coloricon src='data:image/png;base64,{st.session_state.stress_low}'/>
                                 <p>
-                                    """ + translate["low"] + """ <br>
-                                    <span class='""" + class_val + """'>""" + str(stress_duration_low) + """ (""" + str(stress_percentage_low) + """%)</span>
+                                    {translate["low"]} <br>
+                                    <span class='{class_val}'>{str(stress_duration_low)} ({str(stress_percentage_low)}%)</span>
                                 </p>
                             </div>
                             <div class="col-12">
-                                <img class=coloricon src='data:image/png;base64,""" + st.session_state.stress_medium + """'/>
+                                <img class=coloricon src='data:image/png;base64,{st.session_state.stress_medium}'/>
                                 <p>
-                                    """ + translate["medium"] + """ <br>
-                                    <span class='""" + class_val + """'>""" + str(stress_duration_medium) + """ (""" + str(stress_percentage_medium) + """%)</span>
+                                    {translate["medium"]} <br>
+                                    <span class='{class_val}'>{str(stress_duration_medium)} ({str(stress_percentage_medium)}%)</span>
                                 </p>
                             </div>
                             <div class="col-12">
-                                <img class=coloricon src='data:image/png;base64,""" + st.session_state.stress_high + """'/>
+                                <img class=coloricon src='data:image/png;base64,{st.session_state.stress_high}'/>
                                 <p>
-                                    """ + translate["high"] + """ <br>
-                                    <span class='""" + class_val + """'>""" + str(stress_duration_high) + """ (""" + str(stress_percentage_high) + """%)</span>
+                                    {translate["high"]} <br>
+                                    <span class='{class_val}'>{str(stress_duration_high)} ({str(stress_percentage_high)}%)</span>
                                 </p>
                             </div>
                         </div>
                     </div>
                     <div class="col-5">
                         <center>
-                            <img class=donut src='data:image/png;base64,""" + st.session_state.stress_donut + """'/> 
+                            <img class=donut src='data:image/png;base64,{st.session_state.stress_donut}'/> 
                         </center>
                     </div>
                 </div>
             </div>
         </div>
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='overview_health_section'> 
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.sleep_icon + """'/> 
-                        <p class='indicator_name'>""" + translate["sleep"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.sleep_icon}'/> 
+                        <p class='indicator_name'>{translate["sleep"]}</p>
                     </div>
                     <div class="col-7">
                         <div class="row">
                             <div class="col-12">
-                                <img class=coloricon src='data:image/png;base64,""" + st.session_state.sleep_deep + """'/>
+                                <img class=coloricon src='data:image/png;base64,{st.session_state.sleep_deep}'/>
                                 <p>
-                                    """ + translate["deep"] + """ <br>
-                                    <span class='""" + class_val + """'>""" + str(sleep_duration_deep) + """ (""" + str(sleep_percentage_deep) + """%)</span>
+                                    {translate["deep"]} <br>
+                                    <span class='{class_val}'>{str(sleep_duration_deep)} ({str(sleep_percentage_deep)}%)</span>
                                 </p>
                             </div>
                             <div class="col-12">
-                                <img class=coloricon src='data:image/png;base64,""" + st.session_state.sleep_light + """'/>
+                                <img class=coloricon src='data:image/png;base64,{st.session_state.sleep_light}'/>
                                 <p>
-                                    """ + translate["light"] + """ <br>
-                                    <span class='""" + class_val + """'>""" + str(sleep_duration_light) + """ (""" + str(sleep_percentage_light) + """%)</span>
+                                    {translate["light"]} <br>
+                                    <span class='{class_val}'>{str(sleep_duration_light)} ({str(sleep_percentage_light)}%)</span>
                                 </p>
                             </div>
                             <div class="col-12">
-                                <img class=coloricon src='data:image/png;base64,""" + st.session_state.sleep_rem + """'/>
+                                <img class=coloricon src='data:image/png;base64,{st.session_state.sleep_rem}'/>
                                 <p>
-                                    """ + translate["rem"] + """ <br>
-                                    <span class='""" + class_val + """'>""" + str(sleep_duration_rem) + """ (""" + str(sleep_percentage_rem) + """%)</span>
+                                    {translate["rem"]} <br>
+                                    <span class='{class_val}'>{str(sleep_duration_rem)} ({str(sleep_percentage_rem)}%)</span>
                                 </p>
                             </div>
                             <div class="col-12">
-                                <img class=coloricon src='data:image/png;base64,""" + st.session_state.sleep_awake + """'/>
+                                <img class=coloricon src='data:image/png;base64,{st.session_state.sleep_awake}'/>
                                 <p>
-                                    """ + translate["awake"] + """ <br>
-                                    <span class='""" + class_val + """'>""" + str(sleep_duration_awake) + """ (""" + str(sleep_percentage_awake) + """%)</span>  
+                                    {translate["awake"]} <br>
+                                    <span class='{class_val}'>{str(sleep_duration_awake)} ({str(sleep_percentage_awake)}%)</span>  
                                 </p>
                             </div>
                         </div>
                     </div>
                     <div class="col-5">
                         <center>
-                            <img class=donut src='data:image/png;base64,""" + st.session_state.sleep_donut + """'/> 
+                            <img class=donut src='data:image/png;base64,{st.session_state.sleep_donut}'/> 
                         </center>
                     </div>
                 </div>
             </div>
         </div>
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='overview_health_section'> 
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.calories_icon + """'/> 
-                        <p class='indicator_name'>""" + translate["calories"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.calories_icon}'/> 
+                        <p class='indicator_name'>{translate["calories"]}</p>
                     </div>
                     <div class="col-12">
                         <div class="row">
                             <div class="col-12">
-                                <p>""" + translate["calories_total"] + """</p>
+                                <p>{translate["calories_total"]}</p>
                             </div>
                             <div class="col-12">
-                                <p class='total_calories'>""" + str(calories_total) + """ kcals</p>
-                            </div>
-                        </div>
-                    </div>
-                    <div class="col-6">
-                        <div class="row">
-                            <div class="col-12">
-                                <p>""" + translate["calories_active"] + """</p>
-                            </div>
-                            <div class="col-12">
-                                <p class='""" + class_val + """'>""" + str(calories_rest) + """ kcals</p>
+                                <p class='total_calories'>{str(calories_total)} kcals</p>
                             </div>
                         </div>
                     </div>
                     <div class="col-6">
                         <div class="row">
                             <div class="col-12">
-                                <p>""" + translate["calories_resting"] + """</p>
+                                <p>{translate["calories_active"]}</p>
                             </div>
                             <div class="col-12">
-                                <p class='""" + class_val + """'>""" + str(calories_active) + """ kcals</p>
+                                <p class='{class_val}'>{str(calories_rest)} kcals</p>
+                            </div>
+                        </div>
+                    </div>
+                    <div class="col-6">
+                        <div class="row">
+                            <div class="col-12">
+                                <p>{translate["calories_resting"]}</p>
+                            </div>
+                            <div class="col-12">
+                                <p class='{class_val}'>{str(calories_active)} kcals</p>
                             </div>
                         </div>
                     </div>
                 </div>
             </div>
         </div>
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='overview_health_section'> 
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.intensity_icon + """'/> 
-                        <p class='indicator_name'>""" + translate["intensity"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.intensity_icon}'/> 
+                        <p class='indicator_name'>{translate["intensity"]}</p>
                     </div>
                     <div class="col-12">
                         <div class="row">
                             <div class="col-12">
-                                <p>""" + translate["intensity_total"] + """</p>
+                                <p>{translate["intensity_total"]}</p>
                             </div>
                             <div class="col-12">
-                                <p class='total_intensity_minutes'>""" + str(intensity_total) + """ </p>
-                            </div>
-                        </div>
-                    </div>
-                    <div class="col-6">
-                        <div class="row">
-                            <div class="col-12">
-                                <p>""" + translate["intensity_moderate"] + """</p>
-                            </div>
-                            <div class="col-12">
-                                <p class='""" + class_val + """'>""" + str(intensity_moderate) + """ </p>
+                                <p class='total_intensity_minutes'>{str(intensity_total)} </p>
                             </div>
                         </div>
                     </div>
                     <div class="col-6">
                         <div class="row">
                             <div class="col-12">
-                                <p>""" + translate["intensity_vigurous"] + """</p>
+                                <p>{translate["intensity_moderate"]}</p>
                             </div>
                             <div class="col-12">
-                                <p class='""" + class_val + """'>""" + str(intensity_vigurous) + """ </p>
+                                <p class='{class_val}'>{str(intensity_moderate)} </p>
+                            </div>
+                        </div>
+                    </div>
+                    <div class="col-6">
+                        <div class="row">
+                            <div class="col-12">
+                                <p>{translate["intensity_vigurous"]}</p>
+                            </div>
+                            <div class="col-12">
+                                <p class='{class_val}'>{str(intensity_vigurous)} </p>
                             </div>
                         </div>
                     </div>
                 </div>
             </div>
         </div>
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='overview_health_section'> 
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.bodybattery_icon + """'/> 
-                        <p class='indicator_name'>""" + translate["bodybattery"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.bodybattery_icon}'/> 
+                        <p class='indicator_name'>{translate["bodybattery"]}</p>
                     </div>
                     <div class="col-6">
                         <div class="row">
                             <div class="col-12">
-                                <p>""" + translate["high"] + """</p>
+                                <p>{translate["high"]}</p>
                             </div>
                             <div class="col-12">
-                                <p class='body_battery_scores'>""" + str(body_bat_high) + """%</p>
+                                <p class='body_battery_scores'>{str(body_bat_high)}%</p>
                             </div>
                         </div>
                     </div>
                     <div class="col-6">
                         <div class="row">
                             <div class="col-12">
-                                <p>""" + translate["low"] + """</p>
+                                <p>{translate["low"]}</p>
                             </div>
                             <div class="col-12">
-                                <p class='body_battery_scores'>""" + str(body_bat_low) + """%</p>
+                                <p class='body_battery_scores'>{str(body_bat_low)}%</p>
                                 <br>
                             </div>
                         </div>
@@ -777,9 +776,9 @@ def smart_textile_raw_data_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div id="smart_textile_raw_data" class="main_title">
-        <p>""" + translate["smart_textile_raw_data_title"] + """</p>
+        <p>{translate["smart_textile_raw_data_title"]}</p>
     </div>
     """
     
@@ -789,9 +788,9 @@ def smart_textile_raw_data_download():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div>
-        <p>""" + translate["smart_textile_raw_data_download"] + """</p> 
+        <p>{translate["smart_textile_raw_data_download"]}</p> 
     </div>
     """
 
@@ -801,9 +800,9 @@ def health_indicators_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div id="health_indicators" class="main_title">
-        <p>""" + translate["health_indicators"] + """</p>
+        <p>{translate["health_indicators"]}</p>
     </div>
     """
     
@@ -813,9 +812,9 @@ def health_indicators_download():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div>
-        <p>""" + translate["health_indicators_download"] + """</p>
+        <p>{translate["health_indicators_download"]}</p>
     </div>
     """
     
@@ -826,10 +825,10 @@ def health_indicators_heart_bpm_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <img class=icon src='data:image/png;base64,""" + st.session_state.heart_icon + """'/> 
-        <p>""" + translate["health_indicators_heart_bpm_title"] + """</p>
+        <img class=icon src='data:image/png;base64,{st.session_state.heart_icon}'/> 
+        <p>{translate["health_indicators_heart_bpm_title"]}</p>
     </div>
     """
     
@@ -844,14 +843,14 @@ def health_indicators_heart_bpm_results():
     bpm_min     = bpm["min"]
     bpm_max     = bpm["max"]
     
-    html = """
+    html = f"""
     <div class=col1_indicators>
-        <p>""" + translate["median"] + """</p>
-        <p class="indicator_main_value">""" + str(bpm_mean) + """ bpm</p>
-        <p>""" + translate["min"] + """</p>
-        <p class="indicator_value">""" + str(bpm_min) + """ bpm</p>
-        <p>""" + translate["max"] + """</p>
-        <p class="indicator_value">""" + str(bpm_max) + """ bpm</p>
+        <p>{translate["median"]}</p>
+        <p class="indicator_main_value">{str(bpm_mean)} bpm</p>
+        <p>{translate["min"]}</p>
+        <p class="indicator_value">{str(bpm_min)} bpm</p>
+        <p>{translate["max"]}</p>
+        <p class="indicator_value">{str(bpm_max)} bpm</p>
     </div>
     """
     
@@ -861,10 +860,10 @@ def health_indicators_heart_hrv_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <img class=icon src='data:image/png;base64,""" + st.session_state.heart_icon + """'/> 
-        <p>""" + translate["health_indicators_heart_hrv_title"] + """</p>
+        <img class=icon src='data:image/png;base64,{st.session_state.heart_icon}'/> 
+        <p>{translate["health_indicators_heart_hrv_title"]}</p>
     </div>
     """
     
@@ -879,14 +878,14 @@ def health_indicators_heart_hrv_results():
     hrv_min     = hrv["min"]
     hrv_max     = hrv["max"]
     
-    html = """
+    html = f"""
     <div class=col1_indicators>
-        <p>""" + translate["median"] + """</p>
-        <p class="indicator_main_value">""" + str(hrv_mean) + """ ms</p>
-        <p>""" + translate["min"] + """</p>
-        <p class="indicator_value">""" + str(hrv_min) + """ ms</p>
-        <p>""" + translate["max"] + """</p>
-        <p class="indicator_value">""" + str(hrv_max) + """ ms</p>
+        <p>{translate["median"]}</p>
+        <p class="indicator_main_value">{str(hrv_mean)} ms</p>
+        <p>{translate["min"]}</p>
+        <p class="indicator_value">{str(hrv_min)} ms</p>
+        <p>{translate["max"]}</p>
+        <p class="indicator_value">{str(hrv_max)} ms</p>
     </div>
     """
     
@@ -917,88 +916,88 @@ def health_indicators_heart_tachy_brady_qt():
     col_val     = "col-8"
     class_val   = "indicator_value"
     
-    html = """
+    html = f"""
     <div class="row">
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='health_section'>
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.tachycardia_alert_icon + """'/> 
-                        <p class='indicator_name'>""" + translate["tachycardia"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.tachycardia_alert_icon}'/> 
+                        <p class='indicator_name'>{translate["tachycardia"]}</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["hr"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["hr"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(tachy_mean) + """ bpm</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(tachy_mean)} bpm</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["duration"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["duration"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(tachy_duration) + """ min</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(tachy_duration)} min</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["proportion"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["proportion"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(tachy_percentage) + """ %</p>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class='""" + col_section + """'>
-            <div class='health_section'> 
-                <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.bradycardia_alert_icon + """'/> 
-                        <p class='indicator_name'>""" + translate["bradycardia"] + """</p>
-                    </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["hr"] + """</p>
-                    </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(brady_mean) + """ bpm</p>
-                    </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["duration"] + """</p>
-                    </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(brady_duration) + """ min</p>
-                    </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["proportion"] + """</p>
-                    </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(brady_percentage) + """ %</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(tachy_percentage)} %</p>
                     </div>
                 </div>
             </div>
         </div>
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='health_section'> 
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.qt_alert_icon + """'/> 
-                        <p class='indicator_name'>""" + translate["qt_length"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.bradycardia_alert_icon}'/> 
+                        <p class='indicator_name'>{translate["bradycardia"]}</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["median"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["hr"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(qt_mean) + """ ms</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(brady_mean)} bpm</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["min"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["duration"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(qt_min) + """ ms</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(brady_duration)} min</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>""" + translate["max"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["proportion"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(qt_max) + """ ms</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(brady_percentage)} %</p>
+                    </div>
+                </div>
+            </div>
+        </div>
+        <div class='{col_section}'>
+            <div class='health_section'> 
+                <div class="row">
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.qt_alert_icon}'/> 
+                        <p class='indicator_name'>{translate["qt_length"]}</p>
+                    </div>
+                    <div class='{col_key}'>
+                        <p>{translate["median"]}</p>
+                    </div>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(qt_mean)} ms</p>
+                    </div>
+                    <div class='{col_key}'>
+                        <p>{translate["min"]}</p>
+                    </div>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(qt_min)} ms</p>
+                    </div>
+                    <div class='{col_key}'>
+                        <p>{translate["max"]}</p>
+                    </div>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(qt_max)} ms</p>
                     </div>
                 </div>
             </div>
@@ -1012,10 +1011,10 @@ def health_indicators_breath_brpm_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <img class=icon src='data:image/png;base64,""" + st.session_state.breath_icon + """'/> 
-        <p>""" +  translate["health_indicators_breath_brpm_title"] + """</p>
+        <img class=icon src='data:image/png;base64,{st.session_state.breath_icon}'/> 
+        <p>{ translate["health_indicators_breath_brpm_title"]}</p>
     </div>
     """
     
@@ -1030,14 +1029,14 @@ def health_indicators_breath_brpm_results():
     brpm_min     = brpm["min"]
     brpm_max     = brpm["max"]
     
-    html = """
+    html = f"""
     <div class=col1_indicators>
-        <p>""" + translate["median"] + """</p>
-        <p class="indicator_main_value">""" + str(brpm_mean) + """ brpm</p>
-        <p>""" + translate["min"] + """</p>
-        <p class="indicator_value">""" + str(brpm_min) + """ brpm</p>
-        <p>""" + translate["max"] + """</p>
-        <p class="indicator_value">""" + str(brpm_max) + """ brpm</p>
+        <p>{translate["median"]}</p>
+        <p class="indicator_main_value">{str(brpm_mean)} brpm</p>
+        <p>{translate["min"]}</p>
+        <p class="indicator_value">{str(brpm_min)} brpm</p>
+        <p>{translate["max"]}</p>
+        <p class="indicator_value">{str(brpm_max)} brpm</p>
     </div>
     """
     
@@ -1047,10 +1046,10 @@ def health_indicators_breath_brv_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <img class=icon src='data:image/png;base64,""" + st.session_state.breath_icon + """'/> 
-        <p>""" + translate["health_indicators_breath_brv_title"] + """</p>
+        <img class=icon src='data:image/png;base64,{st.session_state.breath_icon}'/> 
+        <p>{translate["health_indicators_breath_brv_title"]}</p>
     </div>
     """
     
@@ -1065,14 +1064,14 @@ def health_indicators_breath_brv_results():
     brv_min     = brv["min"]
     brv_max     = brv["max"]
     
-    html = """
+    html = f"""
     <div class=col1_indicators>
-        <p>""" + translate["median"] + """</p>
-        <p class="indicator_main_value">""" + str(brv_mean) + """ s</p>
-        <p>""" + translate["min"] + """</p>
-        <p class="indicator_value">""" + str(brv_min) + """ s</p>
-        <p>""" + translate["max"] + """</p>
-        <p class="indicator_value">""" + str(brv_max) + """ s</p>
+        <p>{translate["median"]}</p>
+        <p class="indicator_main_value">{str(brv_mean)} s</p>
+        <p>{translate["min"]}</p>
+        <p class="indicator_value">{str(brv_min)} s</p>
+        <p>{translate["max"]}</p>
+        <p class="indicator_value">{str(brv_max)} s</p>
     </div>
     """
     
@@ -1103,88 +1102,88 @@ def health_indicators_breath_tachy_brady_inexratio():
     col_val     = "col-8"
     class_val   = "indicator_value"
     
-    html = """
+    html = f"""
     <div class="row">
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='health_section'> 
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.tachypnea_alert_icon + """'/> 
-                        <p class='indicator_name'>"""+ translate["tachypnea"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.tachypnea_alert_icon}'/> 
+                        <p class='indicator_name'>{translate["tachypnea"]}</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>"""+ translate["br"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["br"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(tachy_mean) + """ brpm</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(tachy_mean)} brpm</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>"""+ translate["duration"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["duration"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(tachy_duration) + """ min</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(tachy_duration)} min</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>"""+ translate["proportion"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["proportion"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(tachy_percentage) + """ %</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(tachy_percentage)} %</p>
                     </div>
                 </div>
             </div>
         </div>
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='health_section'> 
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.tachypnea_alert_icon + """'/> 
-                        <p class='indicator_name'>"""+ translate["bradypnea"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.bradypnea_alert_icon}'/> 
+                        <p class='indicator_name'>{translate["bradypnea"]}</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>"""+ translate["br"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["br"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(brady_mean) + """ brpm</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(brady_mean)} brpm</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>"""+ translate["duration"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["duration"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(brady_duration) + """ min</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(brady_duration)} min</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>"""+ translate["proportion"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["proportion"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(brady_percentage) + """ %</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(brady_percentage)} %</p>
                     </div>
                 </div>
             </div>
         </div>
-        <div class='""" + col_section + """'>
+        <div class='{col_section}'>
             <div class='health_section'> 
                 <div class="row">
-                    <div class='""" + col_title + """'>
-                        <img class=icon src='data:image/png;base64,""" + st.session_state.qt_alert_icon + """'/> 
-                        <p class='indicator_name'>"""+ translate["inout_length_ratio"] + """</p>
+                    <div class='{col_title}'>
+                        <img class=icon src='data:image/png;base64,{st.session_state.qt_alert_icon}'/> 
+                        <p class='indicator_name'>{translate["inout_length_ratio"]}</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>"""+ translate["median"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["median"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(inexratio_mean) + """</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(inexratio_mean)}</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>"""+ translate["min"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["min"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(inexratio_min) + """</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(inexratio_min)}</p>
                     </div>
-                    <div class='""" + col_key + """'>
-                        <p>"""+ translate["max"] + """</p>
+                    <div class='{col_key}'>
+                        <p>{translate["max"]}</p>
                     </div>
-                    <div class='""" + col_val + """'>
-                        <p class='""" + class_val + """'>""" + str(inexratio_max) + """</p>
+                    <div class='{col_val}'>
+                        <p class='{class_val}'>{str(inexratio_max)}</p>
                     </div>
                 </div>
             </div>
@@ -1198,10 +1197,10 @@ def health_indicators_stress_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <img class=icon src='data:image/png;base64,""" + st.session_state.stress_icon + """'/> 
-        <p>""" + translate["stress"] + """</p>
+        <img class=icon src='data:image/png;base64,{st.session_state.stress_icon}'/> 
+        <p>{translate["stress"]}</p>
     </div>
     """
     
@@ -1224,45 +1223,45 @@ def health_indicators_stress_results():
     
     class_val   = "indicator_value"
     
-    html = """
+    html = f"""
     <div class='col-12'>
         <div class='health_section'> 
             <div class="row">
                 <div class="col-7">
                     <div class="row">
                         <div class="col-12">
-                            <img class=coloricon src='data:image/png;base64,""" + st.session_state.stress_rest + """'/> 
+                            <img class=coloricon src='data:image/png;base64,{st.session_state.stress_rest}'/> 
                             <p>
-                                """ + translate["rest"] + """ <br>
-                                <span class='""" + class_val + """'>""" + str(stress_duration_rest) + """  (""" + str(stress_percentage_rest) + """%)</span>
+                                {translate["rest"]} <br>
+                                <span class='{class_val}'>{str(stress_duration_rest)}  ({str(stress_percentage_rest)}%)</span>
                             </p>
                         </div>
                         <div class="col-12">
-                            <img class=coloricon src='data:image/png;base64,""" + st.session_state.stress_low + """'/>
+                            <img class=coloricon src='data:image/png;base64,{st.session_state.stress_low}'/>
                             <p>
-                                """ + translate["low"] + """ <br>
-                                <span class='""" + class_val + """'>""" + str(stress_duration_low) + """  (""" + str(stress_percentage_low) + """%)</span>
+                                {translate["low"]} <br>
+                                <span class='{class_val}'>{str(stress_duration_low)}  ({str(stress_percentage_low)}%)</span>
                             </p>
                         </div>
                         <div class="col-12">
-                            <img class=coloricon src='data:image/png;base64,""" + st.session_state.stress_medium + """'/>
+                            <img class=coloricon src='data:image/png;base64,{st.session_state.stress_medium}'/>
                             <p>
-                                """ + translate["medium"] + """ <br>
-                                <span class='""" + class_val + """'>""" + str(stress_duration_medium) + """  (""" + str(stress_percentage_medium) + """%)</span>
+                                {translate["medium"]} <br>
+                                <span class='{class_val}'>{str(stress_duration_medium)}  ({str(stress_percentage_medium)}%)</span>
                             </p>
                         </div>
                         <div class="col-12">
-                            <img class=coloricon src='data:image/png;base64,""" + st.session_state.stress_high + """'/>
+                            <img class=coloricon src='data:image/png;base64,{st.session_state.stress_high}'/>
                             <p>
-                                """ + translate["high"] + """ <br>
-                                <span class='""" + class_val + """'>""" + str(stress_duration_high) + """  (""" + str(stress_percentage_high) + """%)</span>
+                                {translate["high"]} <br>
+                                <span class='{class_val}'>{str(stress_duration_high)}  ({str(stress_percentage_high)}%)</span>
                             </p>
                         </div>
                     </div>
                 </div>
                 <div class="col-5">
                     <center>
-                        <img class=donut src='data:image/png;base64,""" + st.session_state.stress_donut + """'/> 
+                        <img class=donut src='data:image/png;base64,{st.session_state.stress_donut}'/> 
                     </center>
                 </div>
             </div>
@@ -1276,10 +1275,10 @@ def health_indicators_pulseox_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <img class=icon src='data:image/png;base64,""" + st.session_state.pulseox_icon + """'/> 
-        <p>""" + translate["pulseox"] + """</p>
+        <img class=icon src='data:image/png;base64,{st.session_state.pulseox_icon}'/> 
+        <p>{translate["pulseox"]}</p>
     </div>
     """
     
@@ -1287,22 +1286,22 @@ def health_indicators_pulseox_title():
 
 def health_indicators_pulseox_results():
     
-    html = """
+    html = f"""
     <div class='col-12'>
         <div class='health_section'> 
             <div class="row">
                 <div class="col-6">
-                        <img class=coloricon src='data:image/png;base64,""" + st.session_state.spo2_green + """'/>
+                        <img class=coloricon src='data:image/png;base64,{st.session_state.spo2_green}'/>
                         <p>90 - 100 %</p>
-                        <img class=coloricon src='data:image/png;base64,""" + st.session_state.spo2_yellow + """'/>
+                        <img class=coloricon src='data:image/png;base64,{st.session_state.spo2_yellow}'/>
                         <p>80 - 89 %</p>
-                        <img class=coloricon src='data:image/png;base64,""" + st.session_state.spo2_orange + """'/>
+                        <img class=coloricon src='data:image/png;base64,{st.session_state.spo2_orange}'/>
                         <p>70 - 79 %</p>
-                        <img class=coloricon src='data:image/png;base64,""" + st.session_state.spo2_red + """'/>
+                        <img class=coloricon src='data:image/png;base64,{st.session_state.spo2_red}'/>
                         <p>< 70 %</p>
                 </div>
                 <div class="col-6">
-                    <img class=donut src='data:image/png;base64,""" + st.session_state.spo2_donut + """'/> 
+                    <img class=donut src='data:image/png;base64,{st.session_state.spo2_donut}'/> 
                 </div>
             </div>
         </div>
@@ -1315,10 +1314,10 @@ def health_indicators_bodybattery_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <img class=icon src='data:image/png;base64,""" + st.session_state.bodybattery_icon + """'/> 
-        <p>""" + translate["bodybattery"] + """</p>
+        <img class=icon src='data:image/png;base64,{st.session_state.bodybattery_icon}'/> 
+        <p>{translate["bodybattery"]}</p>
     </div>
     """
     
@@ -1332,12 +1331,12 @@ def health_indicators_bodybattery_results():
     bodybattery_high    = bodybattery["high"]
     bodybattery_low     = bodybattery["low"]
     
-    html = """
+    html = f"""
     <div class=col1_indicators>
-        <p>""" + translate["high"] + """</p>
-        <p class="indicator_main_value">""" + str(bodybattery_high) + """ </p>
-        <p>""" + translate["low"] + """</p>
-        <p class="indicator_value">""" + str(bodybattery_low) + """</p>
+        <p>{translate["high"]}</p>
+        <p class="indicator_main_value">{str(bodybattery_high)} </p>
+        <p>{translate["low"]}</p>
+        <p class="indicator_value">{str(bodybattery_low)}</p>
     </div>
     """
     
@@ -1347,10 +1346,10 @@ def health_indicators_sleep_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <img class=icon src='data:image/png;base64,""" + st.session_state.sleep_icon + """'/> 
-        <p>""" + translate["sleep"] + """</p>
+        <img class=icon src='data:image/png;base64,{st.session_state.sleep_icon}'/> 
+        <p>{translate["sleep"]}</p>
     </div>
     """
     
@@ -1373,45 +1372,45 @@ def health_indicators_sleep_results():
     
     class_val               = "indicator_value"
     
-    html = """
+    html = f"""
     <div class='col-12'>
         <div class='health_section'> 
             <div class="row">
                 <div class="col-7">
                     <div class="row">
                         <div class="col-12">
-                            <img class=coloricon src='data:image/png;base64,""" + st.session_state.sleep_deep + """'/>
+                            <img class=coloricon src='data:image/png;base64,{st.session_state.sleep_deep}'/>
                             <p>
-                                """ + translate["deep"] + """ <br>
-                                <span class='""" + class_val + """'>""" + str(sleep_duration_deep) + """  (""" + str(sleep_percentage_deep) + """%)</span> 
+                                {translate["deep"]} <br>
+                                <span class='{class_val}'>{str(sleep_duration_deep)}  ({str(sleep_percentage_deep)}%)</span> 
                             </p>
                         </div>
                         <div class="col-12">
-                            <img class=coloricon src='data:image/png;base64,""" + st.session_state.sleep_light + """'/>
+                            <img class=coloricon src='data:image/png;base64,{st.session_state.sleep_light}'/>
                             <p>
-                                """ + translate["light"] + """ <br>
-                                <span class='""" + class_val + """'>""" + str(sleep_duration_light) + """  (""" + str(sleep_percentage_light) + """%)</span>
+                                {translate["light"]} <br>
+                                <span class='{class_val}'>{str(sleep_duration_light)}  ({str(sleep_percentage_light)}%)</span>
                             </p>
                         </div>
                         <div class="col-12">
-                            <img class=coloricon src='data:image/png;base64,""" + st.session_state.sleep_rem + """'/>
+                            <img class=coloricon src='data:image/png;base64,{st.session_state.sleep_rem}'/>
                             <p>
-                                """ + translate["rem"] + """ <br>
-                                <span class='""" + class_val + """'>""" + str(sleep_duration_rem) + """  (""" + str(sleep_percentage_rem) + """%)</span>
+                                {translate["rem"]} <br>
+                                <span class='{class_val}'>{str(sleep_duration_rem)}  ({str(sleep_percentage_rem)}%)</span>
                             </p>
                         </div>
                         <div class="col-12">
-                            <img class=coloricon src='data:image/png;base64,""" + st.session_state.sleep_awake + """'/>
+                            <img class=coloricon src='data:image/png;base64,{st.session_state.sleep_awake}'/>
                             <p>
-                                """ + translate["awake"] + """ <br>
-                                <span class='""" + class_val + """'>""" + str(sleep_duration_awake) + """  (""" + str(sleep_percentage_awake) + """%)</span> 
+                                {translate["awake"]} <br>
+                                <span class='{class_val}'>{str(sleep_duration_awake)}  ({str(sleep_percentage_awake)}%)</span> 
                             </p>
                         </div>
                     </div>
                 </div>
                 <div class="col-5">
                     <center>
-                        <img class=donut src='data:image/png;base64,""" + st.session_state.sleep_donut + """'/> 
+                        <img class=donut src='data:image/png;base64,{st.session_state.sleep_donut}'/> 
                     </center>
                 </div>
             </div>
@@ -1425,10 +1424,10 @@ def health_indicators_temperature_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <img class=icon src='data:image/png;base64,""" + st.session_state.temperature_icon + """'/> 
-        <p>""" + translate["health_indicators_temperature_title"] + """</p>
+        <img class=icon src='data:image/png;base64,{st.session_state.temperature_icon}'/> 
+        <p>{translate["health_indicators_temperature_title"]}</p>
     </div>
     """
     
@@ -1443,14 +1442,14 @@ def health_indicators_temperature_results():
     temperature_min     = temperature["min"]
     temperature_max     = temperature["max"]
     
-    html = """
+    html = f"""
     <div class=col1_indicators>
-        <p>""" + translate["median"] + """</p>
-        <p class="indicator_main_value">""" + str(temperature_mean) + """ °C</p>
-        <p>""" + translate["min"] + """</p>
-        <p class="indicator_value">""" + str(temperature_min) + """ °C</p>
-        <p>""" + translate["max"] + """</p>
-        <p class="indicator_value">""" + str(temperature_max) + """ °C</p>
+        <p>{translate["median"]}</p>
+        <p class="indicator_main_value">{str(temperature_mean)} °C</p>
+        <p>{translate["min"]}</p>
+        <p class="indicator_value">{str(temperature_min)} °C</p>
+        <p>{translate["max"]}</p>
+        <p class="indicator_value">{str(temperature_max)} °C</p>
     </div>
     """
     
@@ -1460,9 +1459,9 @@ def data_report_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div id="data_report" class="main_title">
-        <p>""" + translate["data_report"] + """</p>
+        <p>{translate["data_report"]}</p>
     </div>
     """
     
@@ -1473,9 +1472,9 @@ def data_report_download():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div>
-        <p>""" + translate["data_report_download"] + """</p>
+        <p>{translate["data_report_download"]}</p>
     </div>
     """
     
@@ -1485,9 +1484,9 @@ def definitions_title():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div id="definitions" class="main_title">
-        <p>""" + translate["definitions"] + """</p>
+        <p>{translate["definitions"]}</p>
     </div>
     """
     
@@ -1497,30 +1496,30 @@ def definitions_period_and_activity():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <p>""" + translate["period_and_activity"] + """</p>
+        <p>{translate["period_and_activity"]}</p>
     </div>
     <div class='definitions_section'>
         <div class="row">
             <div class="col-lg-5">
                 <div class="col-lg-12">
-                    <img class=definitions_miniicon src='data:image/png;base64,""" + st.session_state.night + """'/> 
-                    <p><b>""" + translate["night"] + """</b>: """ + translate["definition_activity_night"] + """</p>
+                    <img class=definitions_miniicon src='data:image/png;base64,{st.session_state.night}'/> 
+                    <p><b>{translate["night"]}</b>: {translate["definition_activity_night"]}</p>
                 </div>
                 <div class='col-lg-12'>
-                    <img class=definitions_miniicon src='data:image/png;base64,""" + st.session_state.day + """'/> 
-                    <p><b>""" + translate["day"] + """</b>: """ + translate["definition_activity_day"] + """</p>
+                    <img class=definitions_miniicon src='data:image/png;base64,{st.session_state.day}'/> 
+                    <p><b>{translate["day"]}</b>: {translate["definition_activity_day"]}</p>
                 </div>
             </div>
             <div class="col-lg-7">
                 <div class="col-lg-12">
-                    <img class=definitions_miniicon src='data:image/png;base64,""" + st.session_state.rest + """'/> 
-                    <p><b>""" + translate["rest"] + """</b>: """ + translate["definition_activity_rest"] + """</p>
+                    <img class=definitions_miniicon src='data:image/png;base64,{st.session_state.rest}'/> 
+                    <p><b>{translate["rest"]}</b>: {translate["definition_activity_rest"]}</p>
                 </div>
                 <div class="col-lg-12">
-                    <img class=definitions_miniicon src='data:image/png;base64,""" + st.session_state.activity + """'/> 
-                    <p><b>""" + translate["activity"] + """</b>: """ + translate["definition_activity_activity"] + """</p>
+                    <img class=definitions_miniicon src='data:image/png;base64,{st.session_state.activity}'/> 
+                    <p><b>{translate["activity"]}</b>: {translate["definition_activity_activity"]}</p>
                 </div>
             </div>
         </div>
@@ -1533,20 +1532,20 @@ def definitions_alert():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <p>""" + translate["alert"] + """</p>
+        <p>{translate["alert"]}</p>
     </div>
     <div class='definitions_section'>
         <div class="row">
             <div class="col-lg-12">
                 <div class="col-lg-12">
-                    <img class=definitions_miniicon src='data:image/png;base64,""" + st.session_state.alert + """'/> 
-                    <p>""" + translate["definition_alert_red"] + """</p>
+                    <img class=definitions_miniicon src='data:image/png;base64,{st.session_state.alert}'/> 
+                    <p>{translate["definition_alert_red"]}</p>
                 </div>
                 <div class='col-lg-12'>
-                    <img class=definitions_miniicon src='data:image/png;base64,""" + st.session_state.alert_no + """'/> 
-                    <p>""" + translate["definition_alert_green"] + """</p>
+                    <img class=definitions_miniicon src='data:image/png;base64,{st.session_state.alert_no}'/> 
+                    <p>{translate["definition_alert_green"]}</p>
                 </div>
             </div>
         </div>
@@ -1559,52 +1558,52 @@ def definitions_heart():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <p>""" + translate["cardiology"] + """</p>
+        <p>{translate["cardiology"]}</p>
     </div>
     <div class='definitions_section'>
         <div class="row">
             <div class="col-lg-12">
                 <p>
-                    <b>""" + translate["health_indicators_heart_bpm_title"] + """</b>
+                    <b>{translate["health_indicators_heart_bpm_title"]}</b>
                     <br>
-                    """ + translate["definition_bpm"] + """
+                    {translate["definition_bpm"]}
                 </p>
                 <p>
-                    <b>""" + translate["hr_resting"] + """</b>
+                    <b>{translate["hr_resting"]}</b>
                     <br>
-                    """ + translate["definition_hr_resting"] + """
+                    {translate["definition_hr_resting"]}
                 </p>
                 <p>
-                    <b>""" + translate["hr_high"] + """</b>
+                    <b>{translate["hr_high"]}</b>
                     <br>
-                    """ + translate["definition_hr_high"] + """
+                    {translate["definition_hr_high"]}
                 </p>
                 <p>
-                    <b>""" + translate["health_indicators_heart_hrv_title"] + """</b>
+                    <b>{translate["health_indicators_heart_hrv_title"]}</b>
                     <br>
-                    """ + translate["definition_hrv"] + """
+                    {translate["definition_hrv"]}
                 </p>
                 <p>
-                    <b>""" + translate["hrv_resting"] + """</b>
+                    <b>{translate["hrv_resting"]}</b>
                     <br>
-                    """ + translate["definition_hrv_resting"] + """
+                    {translate["definition_hrv_resting"]}
                 </p>
                 <p>
-                    <b>""" + translate["qt_length"] + """</b>
+                    <b>{translate["qt_length"]}</b>
                     <br>
-                    """ + translate["definition_qt_length"] + """
+                    {translate["definition_qt_length"]}
                 </p>
                 <p>
-                    <b>""" + translate["tachycardia"] + """</b>
+                    <b>{translate["tachycardia"]}</b>
                     <br>
-                    """ + translate["definition_tachycardia"] + """
+                    {translate["definition_tachycardia"]}
                 </p>
                 <p>
-                    <b>""" + translate["bradycardia"] + """</b>
+                    <b>{translate["bradycardia"]}</b>
                     <br>
-                    """ + translate["definition_bradycardia"] + """
+                    {translate["definition_bradycardia"]}
                 </p>
             </div>
         </div>
@@ -1617,7 +1616,7 @@ def definitions_breath():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
         <p>Respiratory Indicators</p>
     </div>
@@ -1625,44 +1624,44 @@ def definitions_breath():
         <div class="row">
             <div class="col-lg-12">
                 <p>
-                    <b>""" + translate["health_indicators_breath_brpm_title"] + """</b>
+                    <b>{translate["health_indicators_breath_brpm_title"]}</b>
                     <br>
-                    """ + translate["definition_brpm"] + """
+                    {translate["definition_brpm"]}
                 </p>
                 <p>
-                    <b>""" + translate["br_resting"] + """</b>
+                    <b>{translate["br_resting"]}</b>
                     <br>
-                    """ + translate["definition_br_resting"] + """
+                    {translate["definition_br_resting"]}
                 </p>
                 <p>
-                    <b>""" + translate["br_high"] + """</b>
+                    <b>{translate["br_high"]}</b>
                     <br>
-                    """ + translate["definition_br_high"] + """
+                    {translate["definition_br_high"]}
                 </p>
                 <p>
-                    <b>""" + translate["health_indicators_breath_brv_title"] + """</b>
+                    <b>{translate["health_indicators_breath_brv_title"]}</b>
                     <br>
-                    """ + translate["definition_brv"] + """
+                    {translate["definition_brv"]}
                 </p>
                 <p>
-                    <b>""" + translate["brv_resting"] + """</b>
+                    <b>{translate["brv_resting"]}</b>
                     <br>
-                    """ + translate["definition_brv_resting"] + """
+                    {translate["definition_brv_resting"]}
                 </p>
                 <p>
-                    <b>""" + translate["tachypnea"] + """</b>
+                    <b>{translate["tachypnea"]}</b>
                     <br>
-                    """ + translate["definition_tachypnea"] + """
+                    {translate["definition_tachypnea"]}
                 </p>
                 <p>
-                    <b>""" + translate["bradypnea"] + """</b>
+                    <b>{translate["bradypnea"]}</b>
                     <br>
-                    """ + translate["definition_bradypnea"] + """
+                    {translate["definition_bradypnea"]}
                 </p>
                 <p>
-                    <b>""" + translate["inout_length_ratio"] + """</b>
+                    <b>{translate["inout_length_ratio"]}</b>
                     <br>
-                    """ + translate["definition_inout_length_ratio"] + """
+                    {translate["definition_inout_length_ratio"]}
                 </p>
             </div>
         </div>
@@ -1675,17 +1674,17 @@ def definitions_stress():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <p>""" + translate["stress"] + """</p>
+        <p>{translate["stress"]}</p>
     </div>
     <div class='definitions_section'>
         <div class="row">
             <div class="col-lg-12">
                 <p>
-                    <b>""" + translate["stress_score"] + """</b>
+                    <b>{translate["stress_score"]}</b>
                     <br>
-                    """ + translate["definition_stress_score"] + """
+                    {translate["definition_stress_score"]}
                 </p>
             </div>
         </div>
@@ -1698,17 +1697,17 @@ def definitions_pulseox():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <p>""" + translate["pulseox"] + """</p>
+        <p>{translate["pulseox"]}</p>
     </div>
     <div class='definitions_section'>
         <div class="row">
             <div class="col-lg-12">
                 <p>
-                    <b>""" + translate["spo2"] + """</b>
+                    <b>{translate["spo2"]}</b>
                     <br>
-                    """ + translate["definition_spo2"] + """
+                    {translate["definition_spo2"]}
                 </p>
             </div>
         </div>
@@ -1721,17 +1720,17 @@ def definitions_bodybattery():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <p>""" + translate["bodybattery"] + """</p>
+        <p>{translate["bodybattery"]}</p>
     </div>
     <div class='definitions_section'>
         <div class="row">
             <div class="col-lg-12">
                 <p>
-                    <b>""" + translate["bodybattery"] + """</b>
+                    <b>{translate["bodybattery"]}</b>
                     <br>
-                    """ + translate["definition_bodybattery"] + """
+                    {translate["definition_bodybattery"]}
                 </p>
             </div>
         </div>
@@ -1744,17 +1743,17 @@ def definitions_sleep():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <p>""" + translate["sleep"] + """</p>
+        <p>{translate["sleep"]}</p>
     </div>
     <div class='definitions_section'>
         <div class="row">
             <div class="col-lg-12">
                 <p>
-                    <b>""" + translate["sleep_quality"] + """</b>
+                    <b>{translate["sleep_quality"]}</b>
                     <br>
-                    """ + translate["definition_sleep_quality"] + """
+                    {translate["definition_sleep_quality"]}
                 </p>
             </div>
         </div>
@@ -1767,17 +1766,17 @@ def definitions_temp():
     
     translate = st.session_state.translate
     
-    html = """
+    html = f"""
     <div class="second_title">
-        <p>""" + translate["temperature"] + """</p>
+        <p>{translate["temperature"]}</p>
     </div>
     <div class='definitions_section'>
         <div class="row">
             <div class="col-lg-12">
                 <p>
-                    <b>""" + translate["health_indicators_temperature_title"] + """</b>
+                    <b>{translate["health_indicators_temperature_title"]}</b>
                     <br>
-                    """ + translate["definition_skin_temperature"] + """
+                    {translate["definition_skin_temperature"]}
                 </p>
             </div>
         </div>
